added placeholder labels for bonus percents

+++ view/game_view.py
@@ -20,6 +20,8 @@ from ui.rectangle_progress_bar.money_progress_bar import MoneyProgressBar
 from ui.shader_sprite.game_view_shader_sprite import GameViewShaderSprite
 from ui.label.money_bonus_value_percent_label import MoneyBonusValuePercentLabel
 from ui.label.exp_bonus_value_percent_label import ExpBonusValuePercentLabel
+from ui.label.exp_bonus_placeholder_label import ExpBonusPlaceholderLabel
+from ui.label.money_bonus_placeholder_label import MoneyBonusPlaceholderLabel
 from notifications.exp_bonus_expired_notification import ExpBonusExpiredNotification
 from notifications.money_bonus_expired_notification import MoneyBonusExpiredNotification
 
@@ -71,6 +73,8 @@ class GameView(View):
         self.exp_bonus_multiplier, self.money_bonus_multiplier = USER_DB_CURSOR.fetchone()
         self.exp_bonus_percent_label = ExpBonusValuePercentLabel(parent_viewport=self.viewport)
         self.money_bonus_percent_label = MoneyBonusValuePercentLabel(parent_viewport=self.viewport)
+        self.exp_bonus_placeholder_label = ExpBonusPlaceholderLabel(parent_viewport=self.viewport)
+        self.money_bonus_placeholder_label = MoneyBonusPlaceholderLabel(parent_viewport=self.viewport)
 
     def on_init_content(self):
         CONFIG_DB_CURSOR.execute('SELECT app_width, app_height FROM screen_resolution_config')
@@ -95,12 +99,20 @@ class GameView(View):
         self.exp_progress_bar.on_activate()
         self.money_progress_bar.on_activate()
         if self.exp_bonus_multiplier > 1.0:
+            self.exp_bonus_placeholder_label.delete()
             self.exp_bonus_percent_label.on_update_args((round(self.exp_bonus_multiplier * 100 - 100), ))
             self.exp_bonus_percent_label.create()
+        else:
+            self.exp_bonus_percent_label.delete()
+            self.exp_bonus_placeholder_label.create()
 
         if self.money_bonus_multiplier > 1.0:
+            self.money_bonus_placeholder_label.delete()
             self.money_bonus_percent_label.on_update_args((round(self.money_bonus_multiplier * 100 - 100), ))
             self.money_bonus_percent_label.create()
+        else:
+            self.money_bonus_percent_label.delete()
+            self.money_bonus_placeholder_label.create()
 
         for b in self.buttons:
             if b.to_activate_on_controller_init:
@@ -127,6 +139,8 @@ class GameView(View):
         self.main_clock_label_12h.on_change_screen_resolution(self.screen_resolution)
         self.exp_bonus_percent_label.on_change_screen_resolution(self.screen_resolution)
         self.money_bonus_percent_label.on_change_screen_resolution(self.screen_resolution)
+        self.exp_bonus_placeholder_label.on_change_screen_resolution(self.screen_resolution)
+        self.money_bonus_placeholder_label.on_change_screen_resolution(self.screen_resolution)
         for b in self.buttons:
             b.on_change_screen_resolution(self.screen_resolution)
 
@@ -145,6 +159,8 @@ class GameView(View):
         self.main_clock_label_12h.on_update_opacity(self.opacity)
         self.exp_bonus_percent_label.on_update_opacity(self.opacity)
         self.money_bonus_percent_label.on_update_opacity(self.opacity)
+        self.exp_bonus_placeholder_label.on_update_opacity(self.opacity)
+        self.money_bonus_placeholder_label.on_update_opacity(self.opacity)
         for b in self.buttons:
             b.on_update_opacity(new_opacity)
 
@@ -242,19 +258,25 @@ class GameView(View):
     def on_activate_exp_bonus_code(self, value):
         self.exp_bonus_multiplier = round(1.0 + value, 2)
         if self.is_activated:
+            self.exp_bonus_placeholder_label.delete()
             self.exp_bonus_percent_label.on_update_args((round(self.exp_bonus_multiplier * 100 - 100), ))
             self.exp_bonus_percent_label.create()
 
     def on_deactivate_exp_bonus_code(self):
         self.exp_bonus_multiplier = 1.0
-        self.exp_bonus_percent_label.delete()
+        if self.is_activated:
+            self.exp_bonus_percent_label.delete()
+            self.exp_bonus_placeholder_label.create()
 
     def on_activate_money_bonus_code(self, value):
         self.money_bonus_multiplier = round(1.0 + value, 2)
         if self.is_activated:
+            self.money_bonus_placeholder_label.delete()
             self.money_bonus_percent_label.on_update_args((round(self.money_bonus_multiplier * 100 - 100), ))
             self.money_bonus_percent_label.create()
 
     def on_deactivate_money_bonus_code(self):
         self.money_bonus_multiplier = 1.0
-        self.money_bonus_percent_label.delete()
+        if self.is_activated:
+            self.money_bonus_percent_label.delete()
+            self.money_bonus_placeholder_label.create()
